Fix one password item parsing bug
The Ps field had a float value in it. Easyjson was trying to parse out
an int.

+++ one-password/items/items.go
@@ -12,7 +12,7 @@ type Overview struct {
 	Ainfo string   `json:"ainfo"`
 	Pbe   float64  `json:"pbe"`
 	Pgrng bool     `json:"pgrng"`
-	Ps    int      `json:"ps"`
+	Ps    float64  `json:"ps"`
 	Tags  []string `json:"tags"`
 	Title string   `json:"title"`
 	Url   string   `json:"url"`
